requirements.txt
streamlit
weaviate-client
sentence-transformers
python-dotenv
pymongo
tqdm
